src/emailservice/requirements.txt
flask>=3.0
gunicorn>=21.2
jinja2>=3.1
google-auth>=2.0
python-json-logger>=2.0
pymongo>=4.8
